--- src/components/CheckRole.js
@@ -1,7 +1,0 @@
-import Checkbox from "@mui/material/Checkbox";
-
-function CheckRole() {
-    const label = { inputProps: { "aria-label": "Checkbox demo" } };
-    return <Checkbox {...label} />;
-  }
-export default CheckRole;
