--- src/components/FinishBtn.js
@@ -1,10 +1,0 @@
-import Button from "@mui/material/Button";
-
-function FinishBtn() {
-    return (
-      <Button variant="contained" >
-        送出
-      </Button>
-    );
-  }
-export default FinishBtn;
