--- src/components/LinkRole.js
@@ -1,13 +1,0 @@
-import Link from "@mui/material/Link";
-
-function LinkRole() {
-    return (
-      <div>
-        我已閱讀且同意遵守
-        <Link href="https://monospace.guide/books/manual/page/31fef">
-          新制會員物品管理規範
-        </Link>
-      </div>
-    );
-  }
-  export default LinkRole;
