+++ src/components/MenuBar.js
@@ -1,15 +1,12 @@
 import MenuIcon from "@material-ui/icons/Menu";
-
 import {
-  Box,
   AppBar,
   Toolbar,
   Typography,
   IconButton,
 } from "@material-ui/core";
-function MenuBar() {
-  return (
-    
+function MenuBar(props) {
+  return ( 
       <AppBar class="bar">
         <Toolbar>
           <IconButton
@@ -17,11 +14,10 @@ function MenuBar() {
             edge="start"
             color="inherit"
             aria-label="menu"
-            
           >
             <MenuIcon />
           </IconButton>
-          <Typography variant="h6">鎖櫃登記</Typography>
+          <Typography variant="h6">{props.name}</Typography>
         </Toolbar>
       </AppBar>
   );
