+++ src/components/TextField_phone.js
@@ -1,10 +1,10 @@
 import TextField from "@mui/material/TextField";
-function TextField_phone() {
+function TextField_phone(props) {
     return (
       <TextField
         id="outlined-password-input"
-        label="手機號碼"
-        helperText="請輸入您的手機號碼"
+        label={props.label}
+        helperText={props.helpertext}
       />
     );
   }
